[MDC-205] all: handle missing git worktree for version
During an openwrt build, and sometimes during a tarball build,
it is simply not possible to get the version from the VCS.

In that case, the version *must* have been enshrined into a "version"
file in the top_srcdir.   If it isn't, abort the build.

Make bump-version at the top-level generates a new version file.
Alternatively, rm -f version ; make version will also work.

The version file should be updated, and then tagged.