[MDC-257] simet-ma: include OpenWRT jsonpath as convenience
It is not usual to find OpenWRT's jsonpath (executable: jsonfilter) in
Linux distros, nor its libubox dependency.

Include those as convenience, and when a system version is not found,
build a static version of jsonfilter.  Note that we modify the build
system of both (patches included) as a hack to ensure it will behave in
the way we need it to.

This thing is ugly as heck.  The long term solution is to package
jsonpath linked statically just to libubox, and depend on that at a
package level.