[MDC-245] simet-ma: runner: fix non-posix use of read in templates
POSIX shell "read" requires a variable.  This means "(read ; cat)" to
drop the first line of stdin is not always going to work well... in
fact, it was causing bash itself to complain with "read: 3: arg count".

Use "sed" instead.  We could also have used "tail -n +2", but that is
less portable.