[MDC-76] simet-ma: simet-inetup: tcp stack timeout rework
Set BSD sockets timeouts and TCP user timeouts, and add (commented
because it is currently useless) TCP keep-alive based timeouts, because
we really depend on accurately detecting when packets are not getting
ACK'd, since we currently don't have server->client keepalive messages
at the application protocol level.

Note that some sort of Linux kernel quirk may effectively double the
timeout, but until we manage to track this to an autoritative source or
actual bug report to know if it is enshrined as an kernel-userspace ABI
quirk, we cannot compenaste for it (because if it isn't there, we would
risk incorrectly timing out the connection quite frequently).