[MDC-76] simet-ma: simet-inetup: refactor clocks, add ma_link
Create a reltime() helper function to timestamp events, relative to
machine boot (*not* client runtime).  It must be monotonic and it should
not jump with system time adjustments.

Use reltime() to timestamp every event, and also MA_CONNECT so that we
have a synchronization point.  Use it also for internal timeout clocks,
etc.

Add ma_link/ma_nolink events (not yet hooked to anything), and issue an
ma_link event at connect() time for testing purposes, since we are sure
we have a link if we managed to connect()...