[MDC-76] simet-ma: simet-inetup: add client start/stop events
Add a ma_clientstart event with the timestamp of the client start and
send it after the MA_CONNECT message.

This allows us to track client issues better, and we know everything is
"unknown" between system boot and ma_clientstart, as well as between
ma_clientstop and ma_clientstart.

We have no logic to issue ma_clientstop right now, though, since we
don't have any graceful handling of signals, or a "soft reload" yet.