[MDC-76] simet-ma: simet-inetup: remove system uptime from MA_CONNECT
We will not use the system uptime for anything as far as calculating
the internet availability/uptime is concerned.  Instead, we have the
timestamps from CLOCK_MONOTONIC, and should always use them.

Remove the system uptime from MA_CONNECT allows us to drop the
dependency on sysinfo.h/sysinfo(), which are non-portable.