[MDC-76] simet-ma: simet-inetup: do not storm server on service refusal
The server could refuses service simply by closing the connection, and
we should not reset the backoff timers until we are sure it has actually
accepted to service the client.

Otherwise, we would aggressively try to reconnect should the server drop
the connection after receving our first messages.  In particular, it
will use MA_CONNECT to decide whether it want to talk to us or not, so
this would _always_ happen on several cases.