[MDC-76] simet-ma: simet-inetup: handle SIGTERM, SIGQUIT, SIGHUP
Add signal handling infrastructure, and add logic to cleanly exit on
SIGTERM and SIGQUIT.

SIGINT is not handled on purpose, to avoid the need of SIGKILL in case
of a program bug that results in an infinite loop.

SIGHUP is trapped as a "reload config" function, but not acted upon
(yet), so it will just get ignored.