[MDC-76] simet-ma: simet-inetup: implement clean shutdown sequence
Implement the disconnect, disconnectwait and shutdown states for the
protocol, and issue a "ma_clientstop" message to the server on
disconnect (with a suitable timeout should we fail to deliver).

Trigger the clean shutdown sequence if we get any of the exit signals.

Note that the client *will* spin around with all servers in shutdown
state unless it is told to quit by a signal.