[MDC-299] simet-ma: rework mkdir calls and installdirs
Use AC_PROG_MKDIR_P and $(MKDIR_P), and guard directories with --.

Also, use a more automakish way to create empty dirs.  We do _not_
create <prefix>/var/lock, because creating mode 01777 directories is not
a very good idea, and detecting whether we would need to do it or not
really belongs to runtime.

We shall handle the lock directory creation in the scripts that use it,
but that's a separate commit.