[MDC-198] simet-ma: simet-api: bound simet_register_ma.sh retries
The semantics for simet_register_ma require it to exit, not to keep
trying forever.  And it is useful to be able to ask it to not try too
hard.

Try only a specific numer of times, and add a --boot option to reduce
the number of tries and timeouts.

Also, restart the simet-ma service when not in --boot mode and any
registration information changes.