[MDC-341] simet-ma: debian: enforce auto-updates by cron
Unfortunatelly, unattended-upgrades will refuse to update the packages
if we ever touch a conffile (like /etc/init.d/ scripts!), and that's a
no-go for us.

Add a weekly cron-job that tries to upgrade just simet-ma and its
recommended dependencies (in case we add a new one).  We can tell dpkg
to --force-confmiss --force-confnew in that case.  We also tell ucf to
go into force-conffnew mode, should we ever need to start using it.