[MDC-360] simet-ma: debian: restrict license temporarily
Until the final license is ready, remove any license on modification or
distribution from the entire source, other than debian/ from the notices
sent along with the binaries.

Since we do not distribute the source right (except under NDA) right
now, exactly because of this license issue, that covers the binaries
too.