[MDC-301][MDC-199] simet-ma: handle errors properly
Use a common exit status table (in simet_err.h), for now we have three
copies of it, but we will unify them eventually (MDC-301).
Unfortunately, the code itself is nowhere near as good as it should be
at classifying the errors it gets.

While at it, teach TWAMP to abort on malloc failure (MDC-199), instead
of derreferencing NULL pointers.