[MDC-282] simet-ma: inetup: implement return channel msg handling
Implement a set of tcpaq_ primitives to receive or discard simet uptime2
messages, as well as a generic message dispatcher to process received
messages for the SIMET2 uptime2 protocol.

This is a bit more complex than required to allow direct discards from
the kernel buffer (which allows us to handle messages larger than our
buffer for discarding).  Since we already went that far, walk the extra
mile to not copy twice in userspace, either.