[MDC-282] simet-ma: inetup: same-peer reconnect avoidance
Allow the client to track which measurement peer it was last connected
to.  This is relevant because getaddrinfo() might sort its results, and
always give us the same sequence of peers, and we might want to avoid
[trying to] reconnect to the same peer.

The implementation works well when there are multiple possible IPs to
connect to, which is how SIMET2 was designed to be implemented server-
side.  If there is only one possibility (e.g. only one A and one AAAA
record), the implementation will cause the backoff factor to ramp up
faster.