[MDC-282] simet-ma: inetup: implement server event ma_disconnect
Implement server events parsing, and hook ma_disconnect handling to it.

ma_disconnect is sent by a server (measurement peer) when it is cleanly
shutting down, and also when it is trying to load-balance by migrating
some of its clients to a different server.  In both cases, we want to
try other servers before we attempt to reconnect to the one which asked
us to disconnect.

Note that the server *will* disconnect the client even if the client
ignores an ma_disconnect message.  The message exists so that the client
knows it was not a network failure or protocol failure that caused the
connection to drop.