[MDC-403] simet-ma: simet-runner: add report context task
Teach simet-runner its own canonical name and version, and use that on
the built-in tasks (such as geolocation).  This fixes a buglet in the
geolocation task report, too.

Add a context task (with the agent-info metric) to every report.  We can
extend that task later to have other context metrics.

Assume all simetbox builds are for agent-family "embedded", and all
non-simetbox builds are for agent-family "system-service".

Use lsb_release to get the required operating system name/version data
on non-simetbox builds.