[MDC-424] simet-ma: add a basic Dockerfile for building and running
Add a multi-stage Dockerfile to simet-ma that builds, and allows one to
run the simet-ma client as a dockerized service.  This requires a
non-ancient docker to work.

Add a simet-ma.foreground.sh script (in docker/) to run the services
in foreground.

Note: we don't try very hard to minimize the docker images (i.e. we are
not using alpine linux), and we have a quick hack (install via .deb
package) to ensure we have all runtime dependencies installed in the
runtime container.