[MDC-332] simet-ma: deb: add missing runtime dependencies
Add cron (or any cron-daemon) and gawk or mawk as dependencies.
Container images that try too hard to be minimal might be missing these,
for example.

Add lsb-release as a dependency, there is no reason to assume this one
would be installed when we are already worried about cron or awk...