[MDC-271] simet-ma: simet-runner: enhance for easier services list override
For debugging purposes and also to aid automated calibration setups, one
can set SIMET_SERVICELIST_OVERRIDE in the simet-ma.conf to point to a
JSON file with the list of MPs to be used.

If SIMET_SERVICELIST_OVERRIDE is set in the config file, autorization
token support is disabled.  This means a test setup of measurement peers
that do not validate tokens is necessary for it to be of any use.

An example services.json is included.