[MDC-271] simet-ma: tcpbw: generate token when missing
When doing calibration runs, a special server that does not check
tokens is used.  However, the protocol still requires a token to
be sent, and that token should be unique for the whole duration of
the test.

Generate a random time-based token if one is not specified on the
command line.