[MDC-461] simet-ma: simet-api: contact the controller at lmapd start
Attempt to contact the controller when we (re)start the lmapd daemon,
to fetch a new schedule.  If it fails, we are going to use the old
schedule (or the default schedule, if schedules are not persisted
across reboots).

Don't clobber the downloaded schedule if we race a schedule download.