[MDC-433] simet-ma: simet-api: hardcode report-agent-id to true
RFC8193 and RFC8194 disagree on the default for report-agent-id, and the
reference implementation does something funny.

Since SIMET2 *refuses to deal with reports lacking an agent-id[1], just
hardcode report-agent-id to true, and also have it set on the default
schedule to ensure we detect very early if we start complaining that it
is being set to true twice.

[1] anonymous reports are implemented by changing the agent-id at every
report, through the agent-id and agent token registering process.  The
only thing SIMET2 allows an unregistered, agent-id-less MA to do is to
register itself.  Note that a SIMETBOX might or might not auto-update
its firmware while still unregistered, but that's orthogonal to SIMET2.