[MDC-475] simet-ma: simet-api: add reporting scripts
Add scripts to be called from simet-lmapd to generate and send
reports.

simet_lmap-create-report: calls lmapctl report to render a report, and
compresses the output with gzip.  Unused right now, so we don't install
it.

simet_lmap-send-report: calls lmapctl report to render a report, and
sends the output to the collector.