[MDC-478] simet-ma: simet-api: create stdout wrapper
Create an wrapper script that can encapsulate tool output into a simple
generic "stdout report", and a wrapper for the wrapper script that can
encode into a symlink how to get a version out of a tool, so that we
can use symlinks to enable tools.

Prepare symlink-based tool wrappers ("tasklets") for:
traceroute, traceroute6, ping, ping6, fping, dig, drill, nslookup

(and handle the differences on the BusyBox version of those in OpenWRT).