[MDC-470] simet-ma: simet-api: use /bin/bash where we require it
Makefile already translates the first line from #!/bin/sh to #!/bin/bash
except on OpenWRT builds, but it is best if it can translate *also*
bash to sh when required.

This allows us to mark which scripts require bash-like features such as
set -o pipefail (which are available on OpenWRT's busybox-based sh).