[MDC-433] simet-ma: simet-api: set report-agent-id along with agent-id
Do not force report-agent-id to true without also setting agent-id, as
that precludes simet-lmapd from starting and that can be quite
confusing.

Instead, set it in the same place we set the agent-id.  And do change
detection on the lmap agent-id config file, so that we update it if it
doesn't look exactly like we want it to be (e.g. old one missing the
report-agent-id setting).

Fixes: ae3dd0107739b40c0992dfe297e0ed3f8a77a738
([MDC-433] simet-ma: simet-api: hardcode report-agent-id to true)