[MDC-461] simet-ma: simet-api: ensure we can fetch sched before daemon run
simet*_lmap-fetch-schedule.sh sometimes end up being called before the
initscript for simet-lmapd is given a chance to run.  This is not bad,
it means we would already start with the correct schedule...

However, in that case me must ensure /var/run/lmapd (or whatever is
configured as the base directory for LMAP_SCHEDULE_FILE) has already
been created.