[MDC-480] simet-ma: simet-api: lmap: add agent-info generator
Add a generator for the agent-info ambient metric.  The controller (and
the default schedule) must ensure that one action will be added to each
schedule that has actions with a report destination, generating this
metric.

For now, we don't need to change anything on the defaut schedule because
simet-runner generates it by itself.