[MDC-508] simet-ma: simet-api: properly remember the provider
The geolocation metric requires that the provider used to resolve the
geolocation be informed, and this was broken.  In fact, it was not even
being persisted in cache.

Add this information to the cache, and add some error detection so that
it won't break silently.  Use line 3 for the new information, so that we
don't break ABI with simetbox-openwrt-base::simet_geolocation_legado.sh.

(cherry picked from commit 5bc7ec493ccfccd4b45f1c2424ff95ad3d7aa3fe)