[MDC-637] simet-api: simetbox: use md5sum instead of sha256sum
simetbox_lmap-fetch-schedule.sh uses a checksum utility to compare two
files for equality, due to the lack of a more appropriate tool in
OpenWRT's default profile.

Unfortunately, md5sum was changed to sha256sum during a code
refactoring.  sha256sum is not available (by default) on old OpenWRT,
while md5sum is available on OpenWRT both ancient and new.

Changing it back to md5sum unbreaks the script on older SIMETBOX base
firmwares without the need to enable sha256sum there, and it is
completely transparent to the rest of the system.