[MDC-654] simet-spoofer: run some subtests in parallel
Run subtests in parallel, except for the probe_target test, since it
would become brittle in order to preserve the "at least one must
succeed" logic.

Run three tracefilters in parallel to each destination, since this
is much easier to do than messing with nping to send multiple probe
packets in traceroute mode (*and* much easier to post-process too!).

When starting subtests in parallel, offset each subtest by one second,
this distributes the load a bit, and resulted in better behavior
regarding dropped packets (which are false-positives for filtered
packets).