[MDC-654] simet-spoofer: before/after probe_targets and traceroute
We send non-spoofed probes ("probe_targets" test) at the start and end
of the spoofer test.  At start, we do it to ensure we can actually talk
to the MP and to record the path length.  At the end, we do it again to
give us a change to notice the path has changed.

Name these tests "probe_targets_before" and "probe_targets_after" to
ease post-processing of the report.

Also, add a traceroute at the start ("traceroute_before"), and rename
the other one to "traceroute_after".