[MDC-666] simet-ma: twamp: use appropriately sized types
The internal timestamps in microsseconds MUST be a 64-bit type, since
they have the same epoch as NTP (i.e. they count the microsseconds since
the current NTP epoch).

Also, we must ensure the conversion routines use the larger 64-bit data
types internally, as well.

This fixes the TWAMP reports to match the in-wire absolute time.