[MDC-691] simet-ma: docker: production runtime
New dockerfile to create a reference image for use of simet-ma in
production as a measurement agent running inside a container.  This
image is based on packages downloaded from SIMET instead of built
locally.

This image's dockerfile and helper scripts are hosted in the simet-ma
source so that they are kept in sync, as there is a very strong coupling
between the in-container scripts, configuration files, and the simet-ma
source.

We retain the older Dockerfile that does a local build of simet-ma: it
is very relevant documentation of the build requirements, and very
useful for development.

Switch the Makefile to build and run the new reference image instead of
the "development" docker build (which builds simet-ma from source).

Update the documentation accordingly.

Note that we use Debian stable and its unattented-upgrades facility for
now, and that this is a *heavyweight* container which can do its own
syslogging and maintenance.  We may revisit this decision in the future.