[MDC-691] simet-ma: docker: rework Dockerfile (devel)
Rework the "devel" Dockerfile (build dockerfile) and the helper script
used to prepare a simet-ma container, so that the two dockerfiles (devel
and production) diverge less from each other.

This eases up the work on variants, since they can use the helper script
hooks to modify the behavior of both dockerfiles at once.  It is far
better for future maintenance, that way.

Note that on "devel" mode (local debs are created and installed), the
NIC.br feed for the auto-upgrade system is disabled to ensure the
locally-built simet packages are used by the container.