[MDC-697] simet-ma: simet-api: move pairing-id to persistent storage
Move pairing-id (AGENT_PAIR_FILE) to the simet config dir.  This avoids
a nasty issue with /var/run/simet not exisiting at the time we need to
create that file.

There is not much waste of FLASH write cycles, as we do the usual
"write only if contents changed" dance for AGENT_PAIR_FILE.  Also, the
information contained in this file is not expected to change often.

This fixes quite a nasty issue that would cause simet_register_ma to
abort with an error without running a few of the late hooks, and
causing a memory leak (leaving stuff behind on /tmp).