[MDC-707] simet-ma: lmap: expose twampc to lmap schedule
Expose the twamp client directly to the LMAP scheduler as
"simet.nic.br_twampc".  This will be used by the SIMET-AS LMAP
controller.

Note that it is "twampc", and not "twamp".  "simet.nic.br_twamp" is
reserved for an invocation of the twamp client that uses the SIMET peer
discovery services.