[MDC-708] simet-ma: twamp,tcpbw: output valid JSON
Using JSON array fragments as the output format is very helpful when
using templates to generate reports, since it can trivially be made to
work even when one has to pack several such outputs in a single JSON
table (array) without any extra processing.

It is far less helpful when one wants to directly parse such output as
valid JSON, though.

Add a -r <report mode> parameter to tcpbwc and twampc that changes their
output to be valid JSON (by rendering the array delimiters).

This is required for simet-lmapd compatibility.