[MDC-708] simet-ma: lmap: expose twampc -r1 to lmap engine
Since the default registry is coupled with the default reporting engine,
it is better if the twampc/tcpbwc report mode is considered an internal
matter as far as the controler goes.

Add the "-r 1" command line to the twampc base task.  It can be
overriden if needed by specifying it again (e.g. -r 0), the last one in
the command line "wins".