[MDC-727] simet-ma: tcpbw: more precise test duration
We could undershoot or overshoot the target duration of a test by at
nearly one second per direction.  Use timersub() to ensure the fractions
of a second are taken into account, to reduce this error margin.

The code is still not really precise since it will iterate over all
ready flows before checking the time again.  Also, it depends on how
precise select() timeouts really are.