[MDC-733] docker: ensure correct userid on persistent volume
Docker deployments of simet-ma are strongly encouraged to use persistent
volumes for /opt/simet/etc, and some of the files in there are owned by
"nicbr-simet" or belong to the "nicbr-simet" group.

Unfortunately, during a container update, the uid or gid of nicbr-simet
might change.  And that breaks the measurement engine control plane.

Ensure all files under /opt/simet/etc belong to either root or
nicbr-simet, by chown/chgrp'ing any that don't.