[MDC-735] simet-ma: simet-api: register the SIMETBOX MAC address
Currently, there is no proper way in SIMET2 to map a SIMETBOX MAC
address (label) to an agent-id, nor the inverse.

Unfortunately, the MAC address printed in the SIMETBOX label *is* the
most usable identifier most SIMETBOXes currently have, so we end up
needing this mapping for user support, and to preseed MA groups for
organizations.

Add the MAC address as one of the agent attributes in the Agent
Registration API call done by simet_register_ma.