[MDC-760] simet-ma: twamp: render table to stdout on late failures
After we get to the point that we will attempt to connect to the TWAMP
reflector, render the output table to stdout.

This will allow us to return partial results (or an empty result table)
even if we fail due to some other reason (e.g. reflector refused the
connection, etc).