[MDC-419] simet-ma: twamp: revert per-packet timeout discards
Per-packet timeouts are an OWAMP notion, not present on TWAMP, so we
should not discard a packet that actually got reflected as long as we
received it within our test window.

This partially reverts:
    commit f1a4df029cc91fa730da9b8d062d08f9f7ebd2c9
    [MDC-419] simet-ma: twamp: per-packet timeout, timeouts in us

We leave most of the code in-place but disabled, because we might have
to revisit this.

NOTE: we cannot safely remove the per-packet timeout discards from
clients to be certified according to the "SIMET MA ETP v1 and v1.1",
as the certification requirements are not clear enough on that point.