[MDC-801] simet-api, deb, docker: remove simet-ma_periodic.sh
We have replaced simet-ma_periodic.sh with the lmapd-driven schedule,
so remove it.

For the debian package, we do not remove simet-ma_periodic.sh from
already-installed cron.d/simet-ma, in order to not overwrite local
changes.  Since we removed simet-ma_peroidic.sh, the old entry will be
silently rendered inoperative.