[MDC-724] deb: simet-ma.init: fix status action
Use the correct argument order when calling status_of_proc.  While at
it, document some problems found while debugging the issue.

Fortunately, it fails in such a way that the most interesting cases
(stale or missing pidfile) result in the correct answer as long as
nothing is using the stale PID.