[MDC-804][MDC-810] docker: allow invoke-rc.d simet-lmapd,simet-ma
Add a custom /usr/sbin/invoke-rc.d so that simet package updates will
succeed when they try to reload/stop/start simet-lmapd or simet-ma.

We need this so that package updates will restart the services properly.

As a side-effect, since inetupc is controlled by the simet-ma initscript
and simet_register_ma will reload the simet-ma service when the agent-id
and/or agent-token changes, this also takes care of refreshing those for
inetupc.