[MDC-809] simet-ma: inetup: load agent-id and token from file
Get the agent-id and agent token from a file, instead of directly from
the command line.  This is required to support a "reload" signal, since
we need to somehow get the new values.

Change the command line to accept file paths instead of the values of
the agent-id and agent token.  This breaks compatibility with old
containers.

Update the debian initscript and the docker script to use the new inetup
command line interface.