[MDC-816] simet-api: lmap-wrapper: remove broken test
We were doing a test -x $MYDIR/$MYTOOL, but MYTOOL is empty.
Fortunately, this does test -x in a directory, which always succeed.

OTOH, fixing MYTOOL to TOOL won't fix the logic.  Leave it to
lmap-stdout-wrapper to deal with a non-executable, it is the same
handling it has to do for not-found executables anyway.

So, remove the broken [ -x "$MYDIR/$MYTOOL" ] entirely.