[MDC-833] simet-api: tasklets: implement ICMP traceroute
We find ourselves in need of ICMP traceroute, so implement it.

Add a new helper that can use the traceroute metric name to direct
traceroute to the correct protocol (ICMP, UDP), and use that.

Add icmp-traceroute and udp-traceroute as tasklet tools.  To simplify
the code to implement this, remove the traceroute6 tasklet (it was
useless: we need to be able to force ipv4 just as much as we need to be
able to force ipv6).

Switch to ICMP for the default traceroute tasklet tool: it seems to
survive better, but we will really need to do both to be sure we get at
least one route trace...