[MDC-816] deb: gift CAP_NET_RAW+ep to traceroute
Unfortunately, the Debian packaging for traceroute is old, and either
requires root for ICMP traceroute (traceroute package), or worse, gives
it suid powers (inetutils-traceroute).

Well, that breaks SIMET-MA, so work around it by giving the active
traceroute CAP_NET_RAW capabilities.  We try it when the package is
first installed, and also when starting the service -- in case
traceroute was updated and lost the added capability.