[MDC-843] deb, docker: force ephemeral dir modes and ownership
We must not depend on mkdir -p -m to set the access mode of ephemeral
dirs, in case those directories already exist (mkdir does *nothing* in
that case) with incompatible modes.

This can happen on a package update scenario, for example.

Create the directories with the desired access mode, but chmod them to
that same access mode anyway.  This handle the upgrade scenario where
mkdir -p -m does nothing.