[MDC-848] simet-api: change fping tasklet to exec21ok mode
Our uses of fping typically will result in non-zero exit status, and
very often fping will output to stderr stuff that we wanted to capture
as output.

Switch it to the new exec21ok mode.

In order to do that, change the Makefile.am machinery to allow one to
specify the first field (which used to be <empty>-), which means
everything but fping got a "-" prepended to their definition, and fping
got exec21ok- prepended to its definition.