[MDC-403] simet-runner: report servicelist output as a context metric
Add context metric #64545 support to simet-runner.  This is metric
urn:ietf:metrics:perf:Priv_SPMonitor_Passive_ServiceList-output__Multiple_Raw

This is used by the SIMET team to debug issues, so it is a rather opaque
thing.  We just ship whatever we got from ServiceList, quoted so that it
can be sent as a JSON string.