[MDC-850] simet-runner: fix sending of large reports to collector
The original code was reading the entire report through the command line
to curl (!).  Obviously, this would result in breakage sooner or later,
when the report got too big to fit the command line (or the shell
variable it was using to to that)...

Tell curl to read the file, instead.