[MDC-865] simet-api: simet_register_ma: invalidade creds on status 410
Should the Agent-Registry API call return status 410 ("gone"),
unregister the MA and attempt a new registration.

This happens when the NIC.br servers detect credential cloning, which
typically happens due to the misuse of backups by the user (a full
backup was restored on more than one measurement device).

In some situations, the agent might receive status 410 because an old
backup was restored, even if it were restored just in the correct device
(which is a correct use of backups).  This limits the aggressiveness of
what we can do in the MA when we get a status 410 from Agent-Registry.

We reset the pairing when we remove the unusable credentials (this is
impossible to avoid) but we keep the old pairing information to attempt
pairing recovery.

History-line recovery will be possible only through the vlabel and
pairing.  If there was no pairing in place, past history is lost just
like when an unpaired measurement agent goes through a factory or
identity reset.

A backup of the old credentials (agent-id, token) is kept, under the
"agstat_410" suffix.  Access tokens (which are far more ephemeral) are
just removed without backup.