[MDC-881] inetup: implement (received) DISCONNECT message
Implement a DISCONNECT message (currently unused) that exists internally
in the server code.  If it is received by the client, it will cause a
global disconnection -- disconnect from all servers.

Announce as capability "msg-disconnect" to the server, since previous
versions would ignore such a message.

Note that the client does not send this message to the server at this
time, but the current server code (which was not especially modified to
handle this message) will do exactly that if it receives this message:
disconnect the client (due to a server bug).