[MDC-864] simet-api: simetbox: vlabel anti-cloning support
Track the MAC addresses of the SIMETBox which created a vlabel (and the
firmware version, as it might be needed later for hotfixing or
sysupgrade scripts).

We track all MAC addresses in case a configuration change confuses the
main MAC address detection logic, which is only going to be really
stable on OpenWRT 20.

When no MAC address in the vlabel file matches the main MAC of the
running unit, simet_read_vlabel.sh will abort with an error, and
simet_crete_vlabel.sh will try to create a new vlabel.

If the MAC address heuristic deemed a non-empty vlabel as invalid, *and*
it is different from the new generated one, the old vlabel will be
backed up in case it is later need for some sort of bug repair.

While at it, ensure an umask of 0022 when creating the vlabel file.