[MDC-865] simet-api: simet_register_ma: support cloning detection
Get sideband information from simet_create_vlabel should it detect a
foreign vlabel (i.e. one that does not belong to this device), and call
an overridable function with that information.

By default, the new function will consider the agent-id and token
invalid if the vlabel was flagged as foreign.  Note that this is
currently only implemented on the vlabel layer for SIMETBox: the generic
vlabel layer has no way to detect the foreigness of a vlabel at the
moment.

The SIMETBox openwrt feed package will override the new function and do
its own checks, but it can use the vlabel foreigness hint to help the
heuristics.