[MDC-895] simet-api: add simet_view_results.sh
Add a script (not installed on SIMETBox) that calls the default browser
with the URL for the device results and report generation shared with
SIMETBox.

Set the MA type to "simet-ma".

We have not yet implemented language selection based on
LANG/LANGUAGE/LC_*, it is left for the web API to decide.