[MDC-940] simet-api: add simetbox_fping-to-gateway measurement
Implement simetbox_fping-to-gateway.sh LMAP measurement, which detects
the IPv4 or IPv6 gateway and pings it.

In order to do that, add a LMAPd config.d and registry.d for
simetbox-specific tasks.

The new LMAP task is called "simet.nic.br_gateway-ping".  Its
parameters are: ip4|ip6 to select the protocol, -c # to select the
number of probes (pings).

Should we add the simet-ma counterpart (simet_fping-to-gateway), this
can be reworked to be more transparent.

Note: it needs working ubus, and that requires root on OpenWrt 15.05, or
appropriate ACLs in /usr/share/acl.d on OpenWrt 17.01 and later.  If
ubus is non-functional, it will exit with status 3 (internal error).
However, we do not attempt to detect broken ACLs, just an inaccessible
ubus, for simplicity.