[MDC-951] simet-api: simetbox: include context metrics for ping-to-gateway
We want the geolocation measurement, as well as the agent-info context
metrics on every report.  Add them the straightforward way.

Whie at it, increase the report frequency to reduce latency.

This commit, by itself, is just a stopgap solution: there is no
synchronization between schedules in lmapd unless you enforce it through
event timing.  If the report schedule runs while the ping-to-gateway
schedule is still running, the full report for that ping-to-gateway will
be split over two partial lmap reports.  Which is *very* undesireable.