[MDC-654] simet-spoofer: increase traffic uniqueness
Use a random source and destination port, and add a random 32-byte
payload to the probe traffic.  This *is* used by simet-nping to better
filter the traffic arriving at the echo-server (measurement peer).

Before, we'd use destination ports 53 and 33434 at the same time, but
port 53 is often blocked, while 33434 is used by traceroute and for SIP
traffic of CISCO webex and a few others, and thus it attracts a lot of
grief to any <expletive deleted> transit networks that interfere with
it.

We could still use multiple destination ports, but when we do that,
nping will no longer use the destination port as a key to filter
traffic, and that has proved to be a relevant factor on missing "CAPT"
lines in output (which causes just as much damage as that particular
measureent being blocked).