[MDC-654] simet_spoofer: make it conditional to --enable-simet-spoofer
Add a configure.ac-level switch to enable or disable the inclusion of
the simet_spoofer measurement.

The LMAP configuration is still included, but LMAPd will notice the
"binary" (actually, shell script right now) is missing and remove it
from the capability set when required.  We can revisit this, but it
would either require something like "jq" as a build-time dependency, or
it would complicate the simet-api/config.d and simet-api/registry.d
build tremendously.

Note that simet-spoofer is disabled by default right now, since it is
still preliminary.  If enabled, one must ensure simet-nping is also
available (along with traceroute) otherwise it will fail with an
internal error.