[MDC-959] deb: remove inetutils versions from recommends
The GNU inetutils-traceroute is a bad job with incompatible CLI and
output when compared with the Busybox and "modern linux traceroute"
versions.  It is also pretty much dead upstream.

Remove all inetutils-* from recommends.  It was a mistake to even have
them there, and it caused breakage with our scripts due to the
incompatible CLI.