[MDC-959] lmap-tasks: prefer traceroute.db for tracerouting
Debian and Ubuntu uses their "alternatives" system for traceroute, and a
particular system might have inetutils-traceroute as the active one.

And we don't like inetutils-traceroute's CLI or output.  It is
gratuitously incompatible with the ones that took over the world and
can be found on Busybox, Android, and most distros.

Try to run traceroute.db if it exists, instead of traceroute.  That
fixes it on most situations.