[MDC-959] deb: use versioned dependency on traceroute
We need the traceroute package.  Unfortunately, inetutils-traceroute
"provides" traceroute, thus it already satisfies this dependency should
it already be installed on the system by the time we get installed or
updated, and we end up with an unusable traceroute if "traceroute" isn't
also installed (by luck).

We don't really want to conflict with inetutils-traceroute if we can
help it, as this could get in the way of someone that actually
likes/needs inetutils-traceroute *and* also conceivably cause
unattended-upgrades to skip our updates.

Use a versioned dependency for "traceroute": it won't be satisfied by
the "provides: traceroute" in inetutils-traceroute, and should get the
proper one installed.