[MDC-977] twamp: use CLOCK_MONOTONIC for sensitive timing
The code was using gettimeofday() to sample time and generate TWAMP
timestamps.  This is a Bad Thing, because that clock is the same as
CLOCK_REALTIME, and they are not stable: they can step forwards or
backwards.

Switch to CLOCK_MONOTONIC.  Unfortunately, this clock has an unspecified
origin, and the TWAMP protocol needs absolute timestamps due to its
OWAMP undesired baggage.

So, at the beginning of the TWAMP session, we snapshot CLOCK_REALTIME
and then CLOCK_MONOTONIC to be able to convert from monotonic to
realtime.  We will have a small offset error as we cannot read them at
the same syscall, but that's not going to interfere with the timestamp
delta calculations later.

The monotonic clock is not frequency-stable, but it is the only (fast)
clock we can depend on.  This is the best we can do without the larger
refactoring we need to switch to proper kernel/hardware packet
timestamping.