[MDC-980] simet-api: generate local schedule from templates
Change simet*_lmap-fetch-schedule.sh.in to use templates to create the
local schedule.  The templates themselves will be updated in a separate
commit.

The simet_lmap-fetch-schedule.sh script now requires bash (and its
shellbang line was updated accordingly).  We could avoid adding "local"
to it and thus the known-reason for requiring bash, but we do normally
use extended sh as defined by the intersection of BusyBox sh and bash,
so it is overall safer to just require bash upfront.