[MDC-980] simet-api: detect locally-generated schedules
Add a ma_local tag to one of the schedule entries in the templates for
locally-generated schedules.  The NIC.br LMAP controller won't add or
reflect those.

This let us detect when we are operating with a locally-generated
schedule without the need for separate metadata tracking out-of-line.

Rework simet*_lmap-fetch-schedule to regenerate locally-generated
schedules every time we run.  We already detect when it resulted in the
same schedule, and don't disturb anything else in that case.

This allows for easy updates of locally-injected configuration into the
schedule: just rerun simet*_lmap-fetch-schedule.sh to apply any config
changes (or override it with a new one from the controller).