[MDC-654] simet_spoofer: fetch key and targets from webservice before running
Authenticate to the web service using the agent token, and request the
BCP38 server token (if none was provided in the command line with --key).

Do the same for the probe targets (measurement peers): if none are
specified in the command line, request the list of targets through a web
service using the agent token.

If either information isn't available, abort the measurement with exit
status 22.